requirements.txt
psycopg2-binary
aiogram
openpyxl